Rework ModLauncher classloading hacks to preserve static fields between test runs (#23)
Same hack as before, but using the same `TransformingClassLoader` for every test run. This preserves static fields between test runs, and thus the thread pool from one of the tests is reused... this avoids spamming our TC server with new threads until it decides to fail the run due to excessive resource allocation.