Switch back to ASM for class generation (#1)
* Also benchmark many heterogenous listeners

* Switch back to ASM for class generation

This generates a hidden class, similar to the old ClassLoaderFactory.
However, instead of invoking the method directly, we invoke a constant
method handle. As the class no longer references anything outside of
the bus module, this allows it to be defined using inside the current
module via a MethodHandles.Lookup.

This is slightly slower than using LMF for instance methods, as we need
to do a checked cast on the receiver. However, we can make up for it in
the next commit.

* Convert EventListener to be an abstract class

This ends up being much faster for a large number of heterogenous event
listeners.